backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
bcrypt>=4.0.1
mangum>=0.17.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
